backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
aiofiles==24.1.0
redis==5.2.1
celery==5.4.0
numpy==1.26.4
pillow==11.0.0
pefile==2024.8.26
yara-python==4.5.1
onnxruntime-gpu==1.20.1
onnx==1.17.0
scikit-learn==1.5.2
joblib==1.4.2
mlflow==2.18.0
timm==1.0.12
torch==2.5.1
torchvision==0.20.1
torchcam==0.4.0
python-dotenv==1.0.1
httpx==0.28.1
orjson==3.10.12
